Rename `variables.py' into `fuzz_variables.py' to avoid confusion
