+++ src/fuzz_variables.py
